--- client/src/components/mobile/NavigationMobile.js
@@ -1,110 +1,0 @@
-// @flow
-
-import React, { useState } from 'react';
-import { NavLink } from 'react-router-dom';
-import { MenuOutlined } from '@ant-design/icons';
-import { Drawer, Button, Menu } from 'antd';
-import '../../css_mobile/Navigation.css';
-
-import { useAuth } from '../../utils/use-auth';
-
-const NavMobile = () => {
-  const { authed, authRoleIsEquivalentTo } = useAuth();
-  const [drawerVisible, setDrawerVisible] = useState(false);
-
-  return (
-    <nav>
-      <div className="nav-mobile">
-        <div className="nav-topbar-container">
-          <NavLink exact to="/">
-            <div className="nav-mobile-logo" />
-          </NavLink>
-          <div className="nav-menu-btn-container">
-            <Button
-              onClick={() => setDrawerVisible(true)}
-              block
-              type="link"
-              id="navbar"
-            >
-              <div align="right">
-                <MenuOutlined className="nav-menu-btn-icon" />
-              </div>
-            </Button>
-          </div>
-        </div>
-        <Drawer
-          placement="right"
-          closable={false}
-          onClose={() => setDrawerVisible(false)}
-          visible={drawerVisible}
-        >
-          <Menu
-            className="nav-mobile-menu"
-            mode="inline"
-            onClick={args => {
-              const { key } = args;
-              if (key !== 'resources') {
-                setDrawerVisible(false);
-              }
-            }}
-          >
-            {!authed && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" to="/login">
-                  Login
-                </NavLink>
-              </Menu.Item>
-            )}
-            <Menu.Item className="nav-mobile-menu-item">
-              <NavLink className="nav-mobile-option" exact to="/">
-                Home
-              </NavLink>
-            </Menu.Item>
-            {authed && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" exact to="/saved">
-                  Saved
-                </NavLink>
-              </Menu.Item>
-            )}
-            <Menu.Item className="nav-mobile-menu-item">
-              <NavLink className="nav-mobile-option" exact to="/resources">
-                Resources
-              </NavLink>
-            </Menu.Item>
-            {authRoleIsEquivalentTo('admin') && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" to="/admin">
-                  Admin
-                </NavLink>
-              </Menu.Item>
-            )}
-            {authRoleIsEquivalentTo('admin') && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" to="/edit-home">
-                  Edit Home
-                </NavLink>
-              </Menu.Item>
-            )}
-            {authRoleIsEquivalentTo('admin') && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" to="/role-approval">
-                  Users
-                </NavLink>
-              </Menu.Item>
-            )}
-            {authed && (
-              <Menu.Item className="nav-mobile-menu-item">
-                <NavLink className="nav-mobile-option" to="/logout">
-                  Logout
-                </NavLink>
-              </Menu.Item>
-            )}
-          </Menu>
-        </Drawer>
-      </div>
-    </nav>
-  );
-};
-
-export default NavMobile;
